seperated original sunposition class
