+++ 3DAmsterdam/Assets/Amsterdam3D/Scripts/Sun/SunPosition.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 823b8c4207a1f1c4c993b25c6cd85017
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
